perf(api/og): Module OGImage generation feat(package)!: Update nextra, vercel/og & nextra-theme-docs
BREAKING CHANGES: Nextra loader appear broken but not affect development, need nextjs update to 14v sequentially.
