+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "nextra-proyecto-grand-order",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "Documentación del proyecto grand order.",
   "main": "index.js",
   "scripts": {
@@ -17,13 +17,13 @@
     "@egjs/flicking-plugins": "^4.7.0",
     "@egjs/react-flicking": "^4.10.9",
     "@vercel/analytics": "^1.0.0",
-    "@vercel/og": "^0.5.0",
+    "@vercel/og": "^0.6.1",
     "autoprefixer": "^10.4.13",
     "axios": "^1.3.4",
     "next": "^13.2.1",
     "next-sitemap": "^4.0.2",
-    "nextra": "^2.2.16",
-    "nextra-theme-docs": "^2.2.16",
+    "nextra": "^2.13.2",
+    "nextra-theme-docs": "^2.13.2",
     "postcss": "^8.4.21",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
