+++ utils/domain.js
@@ -3,6 +3,10 @@ export function getHostname () {
         return "proyectograndorder.es"
     }
 
+    if (process.env.NODE_ENV == "development" && !process.env.VERCEL_ENV) {
+        return "localhost:3000"
+    }
+
     return process.env.VERCEL_URL || "proyectograndorder.es"
 }
