+++ utils/index.js
@@ -1,2 +1,3 @@
 export * as search from "./search"
 export * as domain from "./domain"
+export * as random from "./random"
