+++ utils/random.js
@@ -1,0 +1,3 @@
+export default function random(arr) {
+    return arr[Math.floor(Math.random() * arr.length)];
+}
